Add styling for new buttons and change box-shadow property
Added colour changing animation for theme button icons, as well as a width, height, and border-radius. This is so that the button has a circular outline when the user focuses on it via keyboard. Changed values of box-shadow so that the colour is now controlled by a variable to allow it to change depending on which theme is active. Added styling for he new edit and submit buttons for the task items.